Allow anyfile to use the state without errors.
